backend/pyproject.toml
[project]
name = "rag-agent-service"
version = "0.1.0"
description = "RAG Agent Service with OpenAI Agents SDK and FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "openai>=1.3.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.9.0",
    "limits>=3.6.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"